Rewrite mid-level request API
This commit changes said API to use a 'request' object, into which
mixins and parameters can be inserted.  This means we no longer
have a separate function for params and mixins, mixins but no
params, mixins but no params and unmarshalling, etc.

In some cases the new API is already cleaner; in others, it is a
bit more verbose, but is at least consistent and hopefully
understandable.